+++ lib/set_custom_cat_values.dart
@@ -2,23 +2,34 @@ import 'package:flutter/material.dart';
 import 'package:bootstrap_like_modal/bootstrap_like_modal.dart';
 import 'package:bootstrap_typography2/bootstrap_typography2.dart';
 
+/// builds a modal which shows the image preview,
+/// and allows the user to input a cat name and description
 class SetCustomCatValues {
   late final GlobalKey<FormState> _formKey;
   late final BSModal _modal;
   late final TextEditingController _nameController;
   late final TextEditingController _descriptionController;
-  final Function(String name, String description) onSubmit;
+  final Function(String name, String description, bool submitted) onComplete;
+  final Image image;
 
   final String title;
+  late bool _submitted;
 
   SetCustomCatValues({
     required BuildContext context,
-    required this.onSubmit,
+    required this.onComplete,
     this.title = "Set Cat Values:",
+    required this.image,
   }) {
+    // init global key & controllers
     _formKey = GlobalKey<FormState>();
     _nameController = TextEditingController();
     _descriptionController = TextEditingController();
+
+    // say the form has not been submitted
+    _submitted = false;
+
+    // create a modal
     _modal = BSModal(
       bsModalActionsRowOptions: BSModalActionsRowOptions(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -58,10 +69,13 @@ class SetCustomCatValues {
                 return null;
               },
             ),
+            image,
           ],
         ),
       ),
     );
+
+    // add cancel and submit buttons to the actions
     _modal.actions = [
       ElevatedButton(
         style: ButtonStyle(
@@ -82,17 +96,24 @@ class SetCustomCatValues {
         ),
         onPressed: () {
           if (_formKey.currentState!.validate()) {
-            onSubmit(String.fromCharCodes(_nameController.text.codeUnits),
-                String.fromCharCodes(_descriptionController.text.codeUnits));
             _modal.hide();
+            _submitted = true;
           }
         },
       ),
     ];
 
+    // run on complete once the modal closes, note that _submitted will be true
+    // only if the submit button was pressed
+    _modal.onHide = () {
+      onComplete(_nameController.text, _descriptionController.text, _submitted);
+    };
+
+    // show modal once it is created
     _modal.show(context);
   }
 
+  /// a disposal function that should be run when you are done with this class
   void dispose() {
     _descriptionController.dispose();
     _nameController.dispose();